Break main file into multiple files, add Makefile
