+++ main.cpp
@@ -1,0 +1,7 @@
+#include "engine.h"
+
+int main()
+{
+    PlayGame();
+    return 0;
+}
