+++ utils.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+bool isNumber(std::string s);
